12.11.2022 -> improving mapper security. Creating the ability to query and save values from the GIOS api.

+++ src/main/java/pl/kozubek/measuringstation/webClient/MeasuringClient.java
@@ -19,7 +19,7 @@ public class MeasuringClient {
                 MeasuringStationDto[].class));
     }
 
-    public MeasuringDataDto getData(int idStation) {
+    public MeasuringDataDto getData(Long idStation) {
         return callGetMethod("/data/getData/{station}",
                 MeasuringDataDto.class,
                 idStation);
